requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.5.2
asttokens==2.4.0
async-lru==2.0.4
attrs==23.1.0
auto-py-to-exe==2.44.2
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.9.0
blis==1.0.1
bottle==0.13.2
bottle-websocket==0.2.9
catalogue==2.0.10
certifi==2023.7.22
cffi==1.17.1
charset-normalizer==3.2.0
click==8.1.7
cloudpathlib==0.20.0
cohere==5.13.3
colorama==0.4.6
comm==0.1.4
confection==0.1.5
contourpy==1.3.0
cycler==0.12.1
cymem==2.0.10
debugpy==1.7.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
distro==1.9.0
Django==4.1.3
djangorestframework==3.14.0
dnspython==2.7.0
dominate==2.9.1
Eel==0.17.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
executing==1.2.0
fastavro==1.9.7
fastjsonschema==2.18.0
filelock==3.8.0
Flask==3.1.0
Flask-Bootstrap==3.3.7.1
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fonttools==4.53.1
fqdn==1.5.1
fsspec==2024.10.0
future==1.0.0
geographiclib==2.0
geopy==2.4.1
gevent==24.10.3
gevent-websocket==0.10.1
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.26.5
idna==3.4
ipykernel==6.25.2
ipython==8.15.0
ipython-genutils==0.2.0
ipywidgets==8.1.0
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.0
Jinja2==3.1.2
jiter==0.8.2
joblib==1.4.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.24.0
kiwisolver==1.4.5
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.9.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.1
murmurhash==1.0.11
nbclient==0.7.4
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.7
nltk==3.9.1
notebook==7.0.3
notebook_shim==0.2.3
Nuitka==2.4.11
numpy==1.26.4
openai==1.57.3
ordered-set==4.1.0
overrides==7.4.0
packaging==23.1
pandas==2.2.2
pandocfilters==1.5.0
parameterized==0.9.0
parso==0.8.3
pefile==2023.2.7
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==2.5.4
preshed==3.0.9
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psutil==5.9.5
psycopg2==2.9.5
pure-eval==0.2.2
pycparser==2.21
pydantic==2.10.2
pydantic_core==2.27.1
Pygments==2.16.1
pyinstaller==6.11.0
pyinstaller-hooks-contrib==2024.9
pymongo==4.10.1
pyparsing==3.1.4
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2023.3
pywin32==306
pywin32-ctypes==0.2.3
pywinpty==2.0.11
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.0
referencing==0.30.2
regex==2024.11.6
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.10.2
scikit-learn==1.6.0
scipy==1.14.1
seaborn==0.13.2
Send2Trash==1.8.2
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
sniffio==1.3.0
soupsieve==2.5
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.36
sqlparse==0.4.3
srsly==2.4.8
stack-data==0.6.2
terminado==0.17.1
textblob==0.18.0.post0
thinc==8.3.2
threadpoolctl==3.5.0
tinycss2==1.2.1
tokenizers==0.21.0
tornado==6.3.3
tqdm==4.67.0
traitlets==5.9.0
typer==0.14.0
types-requests==2.32.0.20241016
typing_extensions==4.12.2
tzdata==2024.1
uri-template==1.3.0
urllib3==2.0.4
virtualenv==20.16.7
visitor==0.1.3
voila==0.5.2
waitress==3.0.2
wasabi==1.1.3
wcwidth==0.2.6
weasel==0.4.1
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.2
websockets==11.0.3
Werkzeug==3.1.3
whichcraft==0.6.1
widgetsnbextension==4.0.8
wrapt==1.17.0
WTForms==3.2.1
zope.event==5.0
zope.interface==7.1.1
zstandard==0.23.0
